app.py
import streamlit as st
import tensorflow as tf
from PIL import Image

# Your classifier code here
def classify_image(image):
    # Load and preprocess the image, then run your model for classification
    pass

st.title("Deep Learning Classifier")
uploaded_file = st.file_uploader("Choose an image", type="jpg")

if uploaded_file is not None:
    image = Image.open(uploaded_file)
    st.image(image, caption='Uploaded Image', use_column_width=True)
    st.write("Classifying...")
    label = classify_image(image)
    st.write(f"Prediction: {label}")