requirements.txt
streamlit==1.15.2
tensorflow==2.10.0
pillow==9.2.0